requirements.txt
# Browser Tab Screenshot Analyzer Dependencies
# Python 3.11+ required

# Gemini API client
google-genai>=0.1.0

# Schema validation and data modeling
pydantic>=2.0.0

# Image processing
Pillow>=10.0.0
